requirements.txt
img2pdf==0.5.1
Pillow==10.1.0
Requests==2.32.3
